Moved ParseInt/Bool to assemble_cvd
ParseInt was only used there while ParseBool was mostly used there. The
only use of ParseBool outside of assemble_cvd was not using all its
parameters and is better served by calling absl::SimpleAtob directly.
