+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -312,7 +312,6 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:host_info",
         "//cuttlefish/common/libs/utils:network",
-        "//cuttlefish/flag_parser",
         "//cuttlefish/host/commands/assemble_cvd:alloc",
         "//cuttlefish/host/commands/assemble_cvd:assemble_cvd_flags",
         "//cuttlefish/host/commands/assemble_cvd:disk_image_flags_vectorization",
@@ -338,6 +337,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/assemble_cvd/flags:kernel_path",
         "//cuttlefish/host/commands/assemble_cvd/flags:mcu_config_path",
         "//cuttlefish/host/commands/assemble_cvd/flags:memory_mb",
+        "//cuttlefish/host/commands/assemble_cvd/flags:parser",
         "//cuttlefish/host/commands/assemble_cvd/flags:restart_subprocesses",
         "//cuttlefish/host/commands/assemble_cvd/flags:super_image",
         "//cuttlefish/host/commands/assemble_cvd/flags:system_image_dir",
