+++ base/cvd/cuttlefish/host/commands/assemble_cvd/flags.cc
@@ -46,7 +46,7 @@
 #include "cuttlefish/common/libs/utils/container.h"
 #include "cuttlefish/common/libs/utils/contains.h"
 #include "cuttlefish/common/libs/utils/files.h"
-#include "cuttlefish/flag_parser/parser.h"
+#include "cuttlefish/host/commands/assemble_cvd/flags/parser.h"
 #include "cuttlefish/common/libs/utils/host_info.h"
 #include "cuttlefish/common/libs/utils/known_paths.h"
 #include "cuttlefish/common/libs/utils/network.h"
