+++ base/cvd/cuttlefish/host/commands/assemble_cvd/flags/from_gflags.cc
@@ -24,7 +24,7 @@
 #include "absl/strings/str_split.h"
 #include <gflags/gflags.h>
 
-#include "cuttlefish/flag_parser/parser.h"
+#include "cuttlefish/host/commands/assemble_cvd/flags/parser.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
