+++ base/cvd/cuttlefish/host/commands/assemble_cvd/flags/parser.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cuttlefish/flag_parser/parser.h"
+#include "cuttlefish/host/commands/assemble_cvd/flags/parser.h"
 
 #include <string>
 #include <string_view>
